refactor code update links on homepage, not dependent on session anymore update my recipes page, dynamic rendering depending on session now
