+++ src/pages/index.js
@@ -26,7 +26,7 @@ export default function Home() {
             height="150"
           />
           <h2>Hey {session.user.name}!</h2>
-          <button>
+          {/* <button>
             <Link href={`/${session.user.id}/startsession`}>
               Start a Session
             </Link>
@@ -36,6 +36,16 @@ export default function Home() {
           </button>
           <button>
             <Link href={`/${session.user.id}/newrecipe`}>Add new Recipe</Link>
+          </button> */}
+
+          <button>
+            <Link href={`/startsession`}>Start a Session</Link>
+          </button>
+          <button>
+            <Link href={`/myrecipes`}>Show my Recipes</Link>
+          </button>
+          <button>
+            <Link href={`/newrecipe`}>Add new Recipe</Link>
           </button>
           <button
             onClick={() => signOut("google")}
